requirements.txt
asgiref==3.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
cssselect2==0.6.0
Django==4.0.2
django-cors-headers==3.10.0
django-crontab==0.7.1
django-filter==21.1
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-mptt==0.13.4
djangorestframework==3.13.1
easy-thumbnails==2.8.1
gunicorn==20.1.0
idna==3.3
instaloader==4.9.4
lxml==4.9.1
paypalrestsdk==1.13.1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-http-client==3.3.7
pytz==2021.3
reportlab==3.6.11
requests==2.28.1
sendgrid==6.9.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==3.5.0
svglib==1.4.1
tinycss2==1.1.1
tzdata==2021.5
urllib3==1.26.12
webencodings==0.5.1
